fix: board meeting — 8 SQL bugs in agency files
Legal: ob.operator_id → ot.operator_id (no such column on operator_bookings)
Security: users has no deleted_at — remove WHERE clause
Rescue: lat/lng cast to ::float — toFixed() fails on string
Eco: ORDER BY alias::int → ORDER BY expression (alias cast not allowed in ORDER BY)
Quality: booking_booking_status typo → booking_status
Evolution: ORDER BY count::int → ORDER BY COUNT(*) (same alias cast issue)
DevOps: agent_approvals table missing — add .catch() fallback
VibeCoder: ORDER BY fail_count::int → ORDER BY expression

+++ lib/agents/agencies/rescue-agency.ts
@@ -54,8 +54,8 @@ export class RescueAgency {
         SELECT
           id,
           user_id,
-          lat,
-          lng,
+          lat::float,
+          lng::float,
           status,
           created_at::text,
           ROUND(EXTRACT(EPOCH FROM (NOW() - created_at)) / 60)::int AS age_minutes
